+++ vcd/resource_vcd_dnat_test.go
@@ -373,13 +373,17 @@ resource "vcd_dnat" "{{.DnatName}}" {
 }
 `
 const testAccCheckVcdDnatWithExtNetw = `
+data "vcd_external_network" "{{.ExternalNetworkName}}" {
+  name = "{{.ExternalNetworkName}}"
+}
+
 resource "vcd_dnat" "{{.DnatName}}" {
   org             = "{{.Org}}"
   vdc             = "{{.Vdc}}"
-  network_name    = "{{.ExternalNetworkName}}"
+  network_name    = "${data.vcd_external_network.{{.ExternalNetworkName}}.name}"
   network_type    = "ext"
   edge_gateway    = "{{.EdgeGateway}}"
-  external_ip     = "{{.ExternalIp}}"
+  external_ip     = "${data.vcd_external_network.{{.ExternalNetworkName}}.ip_scope[0].static_ip_pool[0].start_address}"
   port            = 7777
   protocol        = "tcp"
   internal_ip     = "10.10.102.60"